refactor(extensions): move NodeCompat to src/extensions/compat/, rename ext-node-compat → ext-node-compatibility
- src/extensions/compat/{node-compat,index}.ts
- src/extensions/interfaces/index.ts shim re-exports from ../compat/
- extensions/ext-node-compatibility (@veryfront/ext-node-compatibility) maps veryfront/extensions/compat
- Root deno.json workspace + ./extensions/compat export
- Consumer stragglers (builtin-extensions.ts, embedding/upload-loader, platform/compat/opaque-deps, runtime name + log labels, JSDoc, tests) updated

Renamed to "node-compatibility" (full word) rather than the category-prefix
form for better self-documentation — the abbreviation "compat" was not
universally obvious. Other categories keep the category-first prefix
pattern (ext-ai-*, ext-cache-*, ext-bundler-*, etc.).
