+++ deno.json
@@ -10,9 +10,9 @@
     "./extensions/ext-auth-jwt",
     "./extensions/ext-bundler-esbuild",
     "./extensions/ext-cache-redis",
-    "./extensions/ext-css-tailwind",
     "./extensions/ext-content-mdx",
-    "./extensions/ext-node-compat",
+    "./extensions/ext-css-tailwind",
+    "./extensions/ext-node-compatibility",
     "./extensions/ext-opentelemetry",
     "./extensions/ext-parser-babel"
   ],
@@ -73,6 +73,7 @@
     "./extensions/auth": "./src/extensions/auth/index.ts",
     "./extensions/bundler": "./src/extensions/bundler/index.ts",
     "./extensions/cache": "./src/extensions/cache/index.ts",
+    "./extensions/compat": "./src/extensions/compat/index.ts",
     "./extensions/content": "./src/extensions/content/index.ts",
     "./extensions/css": "./src/extensions/css/index.ts",
     "./extensions/parser": "./src/extensions/parser/index.ts",
@@ -159,6 +160,7 @@
     "veryfront/testing/assert": "./src/testing/assert.ts",
     "veryfront/testing/bdd": "./src/testing/bdd.ts",
     "veryfront/extensions/bundler": "./src/extensions/bundler/index.ts",
+    "veryfront/extensions/compat": "./src/extensions/compat/index.ts",
     "veryfront/extensions/content": "./src/extensions/content/index.ts",
     "veryfront/extensions/contracts": "./src/extensions/contracts.ts",
     "veryfront/extensions/css": "./src/extensions/css/index.ts",
