+++ src/extensions/compat/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Compat category barrel — Node/Deno compatibility surfaces (sqlite, kreuzberg).
+ *
+ * @module extensions/compat
+ */
+
+export type {
+  KreuzbergExtractor,
+  NodeCompat,
+  NodeCompatSqliteDatabase,
+  SqliteStatement,
+} from "./node-compat.ts";
