+++ src/platform/compat/kv/factory.ts
@@ -4,7 +4,7 @@ import { isDeno } from "../runtime.ts";
 import { MemoryKv } from "./memory-adapter.ts";
 import { SqliteKv } from "./sqlite-adapter.ts";
 import type { Kv, SqliteDatabase } from "./types.ts";
-import type { NodeCompat } from "#veryfront/extensions/interfaces/node-compat.ts";
+import type { NodeCompat } from "#veryfront/extensions/compat/node-compat.ts";
 
 interface GlobalWithDenoKv {
   Deno?: {
@@ -39,7 +39,7 @@ export async function openKv(path?: string): Promise<Kv> {
   } else {
     serverLogger.debug(
       "NodeCompat extension not registered — SQLite KV unavailable. " +
-        "Install @veryfront/ext-node-compat to enable SQLite-backed KV.",
+        "Install @veryfront/ext-node-compatibility to enable SQLite-backed KV.",
     );
   }
